Updated docblocks
Updated docblocks for sanitize to return an array.  The sanitize method is a pass-by-reference array, so I'm not sure if it is supposed to return an array, but that's what it does.  

Updated the container docblock as well because Illuminate\Container\Container would have resolved to Rees\Sanitizer\Illuminate\Container\Container inside the class, which is clearly not intended.  Removed random tabs that were in original file as well.